+++ res/layout-v14/analog_clock_layout.xml
@@ -53,11 +53,20 @@
     </FrameLayout>
 
     <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/pollen_container"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+    <!--
+    Notifications
+    -->
+
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/notification_mediacontrol_bar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_weight="19">
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 
     <include
