+++ res/layout-v14/analog_clock_layout_4.xml
@@ -48,11 +48,20 @@
         android:translationY="0px" />
 
     <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/pollen_container"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+    <!--
+    Notifications
+    -->
+
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/notification_mediacontrol_bar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_weight="19">
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 
     <include
