+++ res/layout-v14/clock_layout.xml
@@ -89,7 +89,6 @@
         android:id="@+id/pollen_container"
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 
     <!--
