+++ res/layout-v14/clock_layout_calendar.xml
@@ -95,15 +95,22 @@
         android:visibility="gone" />
 
     <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/pollen_container"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+    <!--
+        Notifications
+    -->
+
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/notification_mediacontrol_bar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-    <!--
-        Notifications
-    -->
     <include
         layout="@layout/notification_bar"
         android:layout_width="match_parent"
