+++ res/layout-v14/clock_layout_digital.xml
@@ -118,15 +118,21 @@
     </com.google.android.flexbox.FlexboxLayout>
 
     <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/notification_mediacontrol_bar"
+        android:id="@+id/pollen_container"
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 
     <!--
         Notifications
     -->
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/notification_mediacontrol_bar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
     <include
         layout="@layout/notification_bar"
         android:layout_width="match_parent"
