+++ res/layout-v14/clock_layout_digital_flip.xml
@@ -53,15 +53,22 @@
     </LinearLayout>
 
     <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/pollen_container"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+    <!--
+    Notifications
+    -->
+
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/notification_mediacontrol_bar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_weight="19">
-
     </androidx.constraintlayout.widget.ConstraintLayout>
-    <!--
-        Notifications
-    -->
+
     <include
         android:layout_width="match_parent"
         android:layout_height="24dp"
